pyproject.toml
[project]
name = "osint-recon"
version = "0.1.0"
description = "CLI OSINT reconnaissance tool — domain, email & username intel with risk scoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.2",
    "dnspython>=2.8.0",
    "httpx>=0.28.1",
    "pydantic>=2",
    "python-dotenv>=1.2.2",
    "python-whois>=0.9.6",
    "rich>=15.0.0",
    "tldextract>=5.3.1",
]

[project.scripts]
osint-recon = "osint_recon.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
    "ruff>=0.15.10",
]
